requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
python-multipart
numpy
faiss-cpu
langchain
langchain-groq
langchain-text-splitters
langchain-community
langchain-huggingface
sentence-transformers
python-docx
pymupdf
jinja2
streamlit
requests
